# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Styr API
  description: >
    Styr's REST and SSE API: OIDC login, Claude Code sessions, approvals,
    workspaces, profiles and server administration.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: /
security:
  - sessionCookie: []
  - bearerToken: []
paths:
...
  /api/v1/notifications:
    get:
      summary: List notification channels
      operationId: listNotifications
      tags: [notifications]
      responses:
        "200":
          description: Every notification channel (never includes a token)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationChannel"
    post:
      summary: Create a notification channel
      description: Admin only. token, when given, is sealed at rest and never returned again.
      operationId: createNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                kind:
                  type: string
                  enum: [ntfy, webhook]
                name:
                  type: string
                url:
                  type: string
                token:
                  type: string
                events:
                  type: array
                  items:
                    type: string
              required: [kind, name, url]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/notifications/{id}:
    patch:
      summary: Update a notification channel
      description: >
        Admin only. An omitted token leaves it unchanged; an explicit
        empty token clears it.
      operationId: patchNotification
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                kind:
                  type: string
                  enum: [ntfy, webhook]
                name:
                  type: string
                url:
                  type: string
                token:
                  type: string
                events:
                  type: array
                  items:
                    type: string
                enabled:
                  type: boolean
      responses:
        "200":
          description: The updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationChannel"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete a notification channel
      description: Admin only.
      operationId: deleteNotification
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/notifications/{id}/test:
    post:
      summary: Send a test notification
      description: >
        Admin only. Builds a synthetic "test" event and sends it through
        the channel using its decrypted token; the token never appears in
        this (or any) response.
      operationId: testNotification
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The test notification was sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Sending the test notification failed (no detail leaked)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /hooks/{slug}:
    post:
      summary: Inbound webhook delivery
      description: >
        Unauthenticated by session cookie — the credential is the
        trigger's own secret, via header X-Styr-Secret or Authorization
        Bearer, or (for a "github" kind trigger) an X-Hub-Signature-256
        HMAC over the raw body. Not part of /api/v1: no CSRF header
        required, no request timeout. Body is capped at 256 KiB.
      operationId: deliverHook
      security: []
      tags: [hooks]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "202":
          description: Delivery logged (a run may or may not have started, depending on status)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeliveryResult"
        "401":
          description: Bad secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Unknown or disabled trigger
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "413":
          description: Body exceeds the 256 KiB limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: styr_session
    bearerToken:
      type: http
      scheme: bearer
      description: >
        A personal API token (POST /me/api-tokens), sent as `Authorization:
        Bearer styr_pat_...`. Equivalent to sessionCookie for every route,
        except that it is exempt from the X-Requested-With CSRF header
        rule (internal/api/middleware.go's csrfGuard).
  responses:
...
    Forbidden:
      description: The current user may not perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    NotFound:
      description: Not found (or not visible to the current user)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
    UnprocessableEntity:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
          required: [code, message]
      required: [error]
...
    DeliveryResult:
      description: >
        The shape POST /hooks/{slug} and POST /triggers/{id}/test both
        respond with: enough to look the delivery up (or follow it to the
        run it started) without the full Delivery record.
      type: object
      properties:
        delivery_id:
          type: string
        status:
          type: string
          enum: [accepted, deduped, cooldown, storm, rejected, failed, skipped]
        run_id:
          type: string
      required: [delivery_id, status]
...
    NotificationChannel:
      description: Never includes the token — see token_present.
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [ntfy, webhook]
        name:
          type: string
        url:
          type: string
        token_present:
          type: boolean
        events:
          type: array
          items:
            type: string
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
      required: [id, kind, name, url, token_present, events, enabled, created_at]